separated PC and data logger serial streams
used Arduino 1.0's SoftwareSerial library to send all the debugging info
down the USB's serial line to the PC, but send only the final
measurements to the OpenLog device.  ... also created a "Do Nothing"
sketch that essentially wipes the previous sketch from the
microcontroller, so that it doesn't dump a bunch of junk to the OpenLog
while a new sketch is being uploaded.  ... not pretty, but works.